Pacman now eats "food"
Food can be eaten (changing the values inside the 2D array so they change from a 2 (food) to a 1 (empty pathway)). Had a weird issue where the pacman would eat the food, but it'd be deleted from a really weird place. Turns out I was using rows when I was meant to be using columns. lol ya idiot.